Move utilities into schemas.utils package.
