+++ sdks/java/core/src/main/java/org/apache/beam/sdk/schemas/FieldValueSetterFactory.java
@@ -16,23 +16,20 @@
  * limitations under the License.
  */
 
-package org.apache.beam.sdk.values.reflect;
+package org.apache.beam.sdk.schemas;
 
 import java.io.Serializable;
-import org.apache.beam.sdk.annotations.Internal;
+import java.util.List;
+import org.apache.beam.sdk.schemas.FieldValueGetter;
+import org.apache.beam.sdk.schemas.FieldValueSetter;
+import org.apache.beam.sdk.schemas.Schema;
 
-/**
- * <b><i>For internal use only; no backwards-compatibility guarantees.</i></b>
- *
- * <p>An interface to access a field of a class.
- *
- * <p>Implementations of this interface are generated at runtime to map object fields to Row fields.
- */
-@Internal
-public interface FieldValueGetter<ObjectT, ValueT> extends Serializable {
-  ValueT get(ObjectT object);
-
-  String name();
-
-  Class type();
+/** A factory interface for creating {@link FieldValueSetter} objects corresponding to a class. */
+public interface FieldValueSetterFactory extends Serializable {
+  /**
+   * Returns a list of {@link FieldValueGetter}s for the target class.
+   *
+   * <p>The returned list is ordered by the order of matching fields in the schema.
+   */
+  List<FieldValueSetter> createSetters(Class<?> targetClass, Schema schema);
 }
