+++ sdks/java/core/src/main/java/org/apache/beam/sdk/schemas/JavaBeanSchema.java
@@ -22,10 +22,8 @@ import org.apache.beam.sdk.annotations.Experimental;
 import org.apache.beam.sdk.annotations.Experimental.Kind;
 import org.apache.beam.sdk.schemas.utils.JavaBeanUtils;
 import org.apache.beam.sdk.values.TypeDescriptor;
-import org.apache.beam.sdk.values.reflect.FieldValueGetterFactory;
-import org.apache.beam.sdk.values.reflect.FieldValueSetterFactory;
-import org.apache.beam.sdk.values.reflect.JavaBeanGetterFactory;
-import org.apache.beam.sdk.values.reflect.JavaBeanSetterFactory;
+import org.apache.beam.sdk.schemas.utils.JavaBeanGetterFactory;
+import org.apache.beam.sdk.schemas.utils.JavaBeanSetterFactory;
 
 /**
  * A {@link SchemaProvider} for Java Bean objects.
